+++ src/kafka_test.go
@@ -4,13 +4,14 @@ import (
 	"encoding/json"
 	"strconv"
 	"strings"
-	"testing"
 
 	"github.com/newrelic/infra-integrations-sdk/integration"
 	"github.com/newrelic/nri-kafka/src/args"
 	"github.com/newrelic/nri-kafka/src/jmxwrapper"
 	"github.com/newrelic/nri-kafka/src/zookeeper"
 	"github.com/samuel/go-zookeeper/zk"
+	"reflect"
+	"testing"
 )
 
 func benchCoreCollection(b *testing.B, numBrokers, numTopics, numPartitions int) {
@@ -578,4 +579,36 @@ func Benchmark_10b100t100p(b *testing.B) {
 }
 func Benchmark_100b100t100p(b *testing.B) {
 	benchCoreCollection(b, 100, 100, 100)
+}
+
+func Test_enforceTopicLimit(t *testing.T) {
+	overLimit := make([]string, maxTopics+1)
+
+	for i := 0; i < maxTopics+1; i++ {
+		overLimit[i] = "topic"
+	}
+
+	testCases := []struct {
+		name   string
+		topics []string
+		want   []string
+	}{
+		{
+			"Over Limit",
+			overLimit,
+			[]string{},
+		},
+		{
+			"Under Limit",
+			[]string{"topic"},
+			[]string{"topic"},
+		},
+	}
+
+	for _, tc := range testCases {
+		out := enforceTopicLimit(tc.topics)
+		if !reflect.DeepEqual(out, tc.want) {
+			t.Errorf("Test Case %s Failed: Expected '%+v' got '%+v'", tc.name, tc.want, out)
+		}
+	}
 }